de-compose triton kernel registry
